# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "МАХ-Инспектор API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/facility": {
      "post": {
        "tags": [
          "me"
        ],
        "summary": "Create My Facility",
        "operationId": "create_my_facility_api_facility_post",
        "parameters": [
          {
            "name": "x-max-init-data",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Max-Init-Data"
            }
          },
          {
            "name": "x-auth-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Auth-Token"
            }
          },
          {
            "name": "x-start-param",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Start-Param"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/owner/state": {
      "get": {
        "tags": [
          "owner"
        ],
        "summary": "Get State",
        "operationId": "get_state_api_owner_state_get",
        "parameters": [
          {
            "name": "x-max-init-data",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Max-Init-Data"
            }
          },
          {
            "name": "x-auth-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Auth-Token"
            }
          },
          {
            "name": "x-start-param",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Start-Param"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/owner/setup": {
      "put": {
        "tags": [
          "owner"
        ],
        "summary": "Put Setup",
        "operationId": "put_setup_api_owner_setup_put",
        "parameters": [
          {
            "name": "x-max-init-data",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Max-Init-Data"
            }
          },
          {
            "name": "x-auth-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Auth-Token"
            }
          },
          {
            "name": "x-start-param",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Start-Param"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetupRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/owner/settings": {
      "patch": {
        "tags": [
          "owner"
        ],
        "summary": "Patch Settings",
        "operationId": "patch_settings_api_owner_settings_patch",
        "parameters": [
          {
            "name": "x-max-init-data",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Max-Init-Data"
            }
          },
          {
            "name": "x-auth-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Auth-Token"
            }
          },
          {
            "name": "x-start-param",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Start-Param"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/owner/shift-role": {
      "put": {
        "tags": [
          "owner"
        ],
        "summary": "Put Shift Role",
        "operationId": "put_shift_role_api_owner_shift_role_put",
        "parameters": [
          {
            "name": "x-max-init-data",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Max-Init-Data"
            }
          },
          {
            "name": "x-auth-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Auth-Token"
            }
          },
          {
            "name": "x-start-param",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Start-Param"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShiftRoleRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/owner/checkin/reissue": {
      "post": {
        "tags": [
          "owner"
        ],
        "summary": "Reissue Checkin",
        "operationId": "reissue_checkin_api_owner_checkin_reissue_post",
        "parameters": [
          {
            "name": "x-max-init-data",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Max-Init-Data"
            }
          },
          {
            "name": "x-auth-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Auth-Token"
            }
          },
          {
            "name": "x-start-param",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Start-Param"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/owner/facility": {
      "delete": {
        "tags": [
          "owner"
        ],
        "summary": "Delete Facility",
        "operationId": "delete_facility_api_owner_facility_delete",
        "parameters": [
          {
            "name": "x-max-init-data",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Max-Init-Data"
            }
          },
          {
            "name": "x-auth-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Auth-Token"
            }
          },
          {
            "name": "x-start-param",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Start-Param"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/owner/staff": {
      "post": {
        "tags": [
          "owner"
        ],
        "summary": "Post Staff",
        "operationId": "post_staff_api_owner_staff_post",
        "parameters": [
          {
            "name": "x-max-init-data",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Max-Init-Data"
            }
          },
          {
            "name": "x-auth-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Auth-Token"
            }
          },
          {
            "name": "x-start-param",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Start-Param"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewStaff"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/owner/staff/{employee_id}": {
      "patch": {
        "tags": [
          "owner"
        ],
        "summary": "Patch Staff",
        "operationId": "patch_staff_api_owner_staff__employee_id__patch",
        "parameters": [
          {
            "name": "employee_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Employee Id"
            }
          },
          {
            "name": "x-max-init-data",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Max-Init-Data"
            }
          },
          {
            "name": "x-auth-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Auth-Token"
            }
          },
          {
            "name": "x-start-param",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Start-Param"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StaffUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "owner"
        ],
        "summary": "Delete Staff",
        "operationId": "delete_staff_api_owner_staff__employee_id__delete",
        "parameters": [
          {
            "name": "employee_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Employee Id"
            }
          },
          {
            "name": "x-max-init-data",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Max-Init-Data"
            }
          },
          {
            "name": "x-auth-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Auth-Token"
            }
          },
          {
            "name": "x-start-param",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Start-Param"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SetupRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "title": "Name"
          },
          "address": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 300
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "qr_checkin": {
            "type": "boolean",
            "title": "Qr Checkin",
            "default": false
          },
          "positions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Positions",
            "default": []
          },
          "features": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Features",
            "default": []
          },
          "assignments": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Assignments",
            "default": {}
          },
          "custom_duties": {
            "items": {
              "$ref": "#/components/schemas/CustomDuty"
            },
            "type": "array",
            "title": "Custom Duties",
            "default": []
          },
          "new_staff": {
            "items": {
              "$ref": "#/components/schemas/NewStaff"
            },
            "type": "array",
            "title": "New Staff",
            "default": []
          },
          "owner_works_shift": {
            "type": "boolean",
            "title": "Owner Works Shift",
            "default": false
          },
          "owner_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Name"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "SetupRequest"
      },
      "ShiftRoleRequest": {
        "properties": {
          "works": {
            "type": "boolean",
            "title": "Works"
          },
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "works"
        ],
        "title": "ShiftRoleRequest"
      },
...
    }
  }
}